05-big-data-avanzado/06-data-lake/proyecto-01-delta-lake/requirements.txt
delta-spark>=2.4.0
pyspark>=3.5.0
